requirements.txt
fastapi
uvicorn
requests
PyPDF2
python-docx
pydantic
google-generativeai
sentence-transformers
redis
python-multipart
python-json-logger
structlog
aiofiles 